Track Controller: Extend query params and remove track routes in other controllers
