+++ server/src/track/track.controller.ts
@@ -1,10 +1,12 @@
 import {
-	Body, Controller, Get, Inject, Post, Put, forwardRef
+	Body, Controller, Get, Inject, Post, Put, Query, forwardRef
 } from '@nestjs/common';
 import { PaginationParameters } from 'src/pagination/models/pagination-parameters';
 import TrackQueryParameters from './models/track.query-parameters';
 import TrackService from './track.service';
-import { ApiOperation, ApiTags } from '@nestjs/swagger';
+import {
+	ApiOperation, ApiPropertyOptional, ApiTags, IntersectionType
+} from '@nestjs/swagger';
 import { TrackType } from '@prisma/client';
 import ReassignTrackDTO from './models/reassign-track.dto';
 import { TrackResponseBuilder } from './models/track.response';
@@ -15,6 +17,44 @@ import Admin from 'src/roles/admin.decorator';
 import IdentifierParam from 'src/identifier/identifier.pipe';
 import Response, { ResponseType } from 'src/response/response.decorator';
 import SongService from 'src/song/song.service';
+import { IsEnum, IsOptional } from 'class-validator';
+import TransformIdentifier from 'src/identifier/identifier.transform';
+import LibraryService from 'src/library/library.service';
+import LibraryQueryParameters from 'src/library/models/library.query-parameters';
+import ReleaseQueryParameters from 'src/release/models/release.query-parameters';
+import ReleaseService from 'src/release/release.service';
+import SongQueryParameters from 'src/song/models/song.query-params';
+
+class Selector extends IntersectionType(TrackQueryParameters.SortingParameter) {
+	@IsOptional()
+	@ApiPropertyOptional({
+		description: 'Filter tracks by library'
+	})
+	@TransformIdentifier(LibraryService)
+	library?: LibraryQueryParameters.WhereInput;
+
+	@IsOptional()
+	@ApiPropertyOptional({
+		description: 'Filter tracks by release'
+	})
+	@TransformIdentifier(ReleaseService)
+	release?: ReleaseQueryParameters.WhereInput;
+
+	@IsOptional()
+	@ApiPropertyOptional({
+		description: 'Filter tracks by song'
+	})
+	@TransformIdentifier(SongService)
+	song?: SongQueryParameters.WhereInput;
+
+	@IsOptional()
+	@ApiPropertyOptional({
+		enum: TrackType,
+		description: 'Filter tracks by type'
+	})
+	@IsEnum(TrackType)
+	type?: TrackType;
+}
 
 @ApiTags("Tracks")
 @Controller('tracks')
@@ -27,7 +67,7 @@ export class TrackController {
 	) { }
 
 	@ApiOperation({
-		summary: 'Get all tracks'
+		summary: 'Get many tracks'
 	})
 	@Get()
 	@Response({
@@ -35,15 +75,14 @@ export class TrackController {
 		type: ResponseType.Page
 	})
 	async getMany(
+		@Query() selector: Selector,
 		@PaginationQuery()
 		paginationParameters: PaginationParameters,
 		@RelationIncludeQuery(TrackQueryParameters.AvailableAtomicIncludes)
-		include: TrackQueryParameters.RelationInclude,
-		@SortingQuery(TrackQueryParameters.SortingKeys)
-		sortingParameter: TrackQueryParameters.SortingParameter
+		include: TrackQueryParameters.RelationInclude
 	) {
 		return this.trackService.getMany(
-			{}, paginationParameters, include, sortingParameter
+			selector, paginationParameters, include, selector
 		);
 	}
